wip: Reorganize existing classes, implement and cleanup
