+++ packages/dev/core/src/Audio/v2/staticSound.ts
@@ -1,15 +1,15 @@
 /* eslint-disable babylonjs/available */
 /* eslint-disable jsdoc/require-jsdoc */
 
-import type { IAudioEngine, IStaticSoundOptions } from "./audioEngine";
-import { AbstractSound } from "./sound";
-import type { ISound } from "./sound";
+import type { IAudioEngine } from "./abstractAudioEngine";
+import type { ISound, IStaticSoundOptions } from "./abstractSound";
+import { AbstractSound } from "./abstractSound";
 import type { IVirtualVoice } from "./virtualVoice";
 import { VirtualVoice, VirtualVoiceType } from "./virtualVoice";
 
 export class StaticSound extends AbstractSound implements ISound {
-    public constructor(audioEngine: IAudioEngine, options: IStaticSoundOptions) {
-        super(audioEngine, options);
+    public constructor(options?: IStaticSoundOptions, audioEngine?: IAudioEngine) {
+        super(options, audioEngine);
 
         this._sourceId = this.audioEngine.physicalEngine.createBuffer(options);
     }
