+++ packages/dev/core/src/Audio/v2/streamingSound.ts
@@ -1,15 +1,15 @@
 /* eslint-disable babylonjs/available */
 /* eslint-disable jsdoc/require-jsdoc */
 
-import type { IAudioEngine, IStreamingSoundOptions } from "./audioEngine";
-import type { ISound } from "./sound";
-import { AbstractSound } from "./sound";
+import type { IAudioEngine } from "./abstractAudioEngine";
+import type { ISound, IStreamingSoundOptions } from "./abstractSound";
+import { AbstractSound } from "./abstractSound";
 import type { IVirtualVoice } from "./virtualVoice";
 import { VirtualVoice, VirtualVoiceType } from "./virtualVoice";
 
 export class StreamingSound extends AbstractSound implements ISound {
-    public constructor(audioEngine: IAudioEngine, options: IStreamingSoundOptions) {
-        super(audioEngine, options);
+    public constructor(options?: IStreamingSoundOptions, audioEngine?: IAudioEngine) {
+        super(options, audioEngine);
 
         this._sourceId = this.audioEngine.physicalEngine.createStream(options);
     }
